% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_apikey}
\alias{check_apikey}
\title{Check if the Environment Variable for the Esports API is set}
\usage{
check_apikey()
}
\description{
`check_apikey` informs the user whether the LoL Esports API Key is available as an environmental variable.
}
