% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandings.R
\name{create_match_df}
\alias{create_match_df}
\title{create match dataframe object}
\usage{
create_match_df(list_of_matches)
}
\arguments{
\item{list_of_matches}{list of all the matches from a given tournament}
}
\description{
Unpacks the list of teams in a match list into a single row data frame
}
