% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_window.R
\name{get_window}
\alias{get_window}
\title{Get game data from a certain starting time}
\usage{
get_window(gameId, hl = "en-US", startingTime = NULL, save_details = FALSE)
}
\arguments{
\item{gameId}{string. The ID of the game for which data should be searched.}

\item{hl}{string. Locale or language code using ISO 639-1 and ISO 3166-1 alpha-2.}

\item{startingTime}{a string in the form `Y-m-d H:M:S`. Needs to be compatible with
`lubridate::ymd_hms()`.}

\item{save_details}{logical. Should details be saved?}
}
\value{
A list of game metadata and a data.frame of game data, or raw query result.
}
\description{
The function returns player data in a specified timeframe. It is possible
to set a certain starting point for the timeframe, but final length is not
able to be predetermined.
}
