% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_completed_events.R
\name{get_completed_events}
\alias{get_completed_events}
\title{Get a list of completed events}
\usage{
get_completed_events(tournamentId, save_details = FALSE, hl = "en-US")
}
\arguments{
\item{tournamentId}{string. Can be taken from the get_leagues() function.}

\item{save_details}{logical. Should just a table be returned, or also details about the requests.}

\item{hl}{string. Locale or language code using ISO 639-1 and ISO 3166-1 alpha-2.}
}
\value{
Returns something
}
\description{
This function takes a tournament Id and returns completed events
}
