% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGames.R
\name{getGames}
\alias{getGames}
\title{getGames}
\usage{
getGames(gameId, hl = "en-US")
}
\arguments{
\item{gameId}{The id of the game of interest}

\item{hl}{string. Locale or language code using ISO 639-1 and ISO 3166-1 alpha-2.}
}
\description{
getGames
}
