% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{getLeagues}
\alias{getLeagues}
\title{Get information about a league, the standings and its various tournaments.}
\usage{
getLeagues(save_details = FALSE, hl = "en-US")
}
\arguments{
\item{save_details}{logical. Should a detailed list, including the API
response be returned, or just a data.frame?}

\item{hl}{string. Locale or language code using ISO 639-1 and ISO 3166-1 alpha-2.}
}
\value{
A list containing a data.frame witht the information about the leagues,
        the used language code and the response to the original GET request
        contained in the function.
}
\description{
The function is used to get detailed information about specific leagues.
The data.frame contained in the returned list gives information about
a leagues region, id, icon and more.
}
