% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLive.R
\name{getLive}
\alias{getLive}
\title{Get live events}
\usage{
getLive(save_details = FALSE, hl = "en-US")
}
\arguments{
\item{save_details}{logical. Should just a table be returned, or also details about the requests.}

\item{hl}{string. Locale or language code using ISO 639-1 and ISO 3166-1 alpha-2.}
}
\value{
Returns an object of class `live` if details were saved. Otherwise, a `data.frame`
        of current live events for the all E-Sports leagues included in the E-Sports API.
}
\description{
Get information about live events currently going on.
}
