% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSchedule.R
\name{getSchedule}
\alias{getSchedule}
\title{Get Information about current Schedules}
\usage{
getSchedule(
  leagueId,
  check_old_pages = TRUE,
  pageToken = NULL,
  save_details = FALSE,
  hl = "en-US"
)
}
\arguments{
\item{leagueId}{string. The league id to be queried.}

\item{check_old_pages}{logical. Should older pages be querried as well?}

\item{pageToken}{Base 64 encoded string used to determine the next "page" of data to pull.
Only used if `check_old_pages` is `TRUE`.}

\item{save_details}{logical. Shoudl details be saved?}

\item{hl}{string. Locale or language code using ISO 639-1 and ISO 3166-1 alpha-2.}
}
\value{
Returns an object of class `Schedule` if details were saved. Otherwise, a `data.frame`
        of scheduled matches for the queried league.
}
\description{
For a given `leagueId` string, the function returns schedule data.
}
