% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandings.R
\name{getStandings}
\alias{getStandings}
\title{Query the standings of a specified tournament}
\usage{
getStandings(tournamentId, save_details = FALSE, hl = "en-US")
}
\arguments{
\item{tournamentId}{string. Can be taken from the getLeagues() function.}

\item{save_details}{logical. Should just a table be returned, or also details about the requests.}

\item{hl}{string. Locale or language code using ISO 639-1 and ISO 3166-1 alpha-2.}
}
\value{
Returns something
}
\description{
The function takes a tournamend ID as argument and returns either a data.frame of the current standings
for the selected tournament, or a list including the data.frame as an element, plus the raw API response.
}
