% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeams.R
\name{getTeams}
\alias{getTeams}
\title{Get information about a team and its roster}
\usage{
getTeams(id, save_details = FALSE, hl = "en-US")
}
\arguments{
\item{id}{string. The}

\item{save_details}{logical.}

\item{hl}{string. Locale or language code using ISO 639-1 and ISO 3166-1 alpha-2.}
}
\description{
Get information about a team and its roster
}
