% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{getTournamentsForLeague}
\alias{getTournamentsForLeague}
\title{Get Tournaments for a given league.}
\usage{
getTournamentsForLeague(leagueId, save_details = FALSE, hl = "en-US")
}
\arguments{
\item{leagueId}{string. The id of the league you want details of.}

\item{save_details}{logical. Should a detailed list, including the API response be returned, or just a data.frame?}

\item{hl}{string. Locale or language code using ISO 639-1 and ISO 3166-1 alpha-2.}
}
\value{
list.
}
\description{
A detailed description of the function?
}
