% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standings.R
\name{process_matches}
\alias{process_matches}
\title{Processes matches}
\usage{
process_matches(parsed)
}
\arguments{
\item{parsed}{parsed data from a get request}
}
\description{
processes the parsed standings data
}
