% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandings.R
\name{process_standings}
\alias{process_standings}
\title{Process Standings Data}
\usage{
process_standings(parsed)
}
\arguments{
\item{parsed}{list. The parsed data from the standings GET request.}
}
\description{
Parses the Data about current standings in the respective league into an
easy to handle `data.frame`.
}
