% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_api}
\alias{query_api}
\title{Query function}
\usage{
query_api(url, key, ...)
}
\arguments{
\item{url}{the url that needs to be queries}

\item{key}{api key}

\item{...}{Possible query parameters}
}
\description{
Builds a GET Query for the LOL Esports API and returns
a list with response, status code and parsed data
}
