% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_apikey}
\alias{set_apikey}
\title{Set an Environment Variable for the public Esports API key}
\usage{
set_apikey()
}
\description{
`set_apikey` creates an environmental variable for the League of Legend Esports API Key,
to be used by other functions calling the API and requiring authentification.
}
