% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventDetails.R
\name{getEventDetails}
\alias{getEventDetails}
\title{getEventDetails}
\usage{
getEventDetails(matchId, hl = "en-US")
}
\arguments{
\item{matchId}{the id of the match of interest.}

\item{hl}{string. Locale or language code using ISO 639-1 and ISO 3166-1 alpha-2.}
}
\description{
getEventDetails
}
