% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventDetails.R
\name{getEventDetails}
\alias{getEventDetails}
\title{getEventDetails}
\usage{
getEventDetails(matchId, hl = "en-US", save_details = FALSE)
}
\arguments{
\item{matchId}{the id of the match of interest.}

\item{hl}{string. Locale or language code using ISO 639-1 and ISO 3166-1 alpha-2.}

\item{save_details}{logical. Should just a table be returned, or also details about the requests.}
}
\value{
data.frame of event details or raw query.
}
\description{
Get details for a given match ID. This is useful to get the
`gameId`s of single games for a given match. So one match can
have different `gameId`s, which can be found by `value$games$game_id`
where `value` is the response value of this function with the `save_details`
flag set to false.
These `gameId`s can be used for further analysis in the functions `getWindow()`
and `getDetails()`.
}
